Basic game commit
This is the commit for the basic game, all done in one go. Reads user
input rock/paper/scissors over 5 rounds and compares to the computer.
Each round it alerts the winner in a popup and after the 5 rounds it
alerts the winner of the match in a popup as well.